feat: add built in 'clock' function
