+++ lox/builtins.ts
@@ -1,0 +1,15 @@
+import { LoxCallable, LoxReturnValue } from "./internal";
+
+export class ClockBuiltin extends LoxCallable {
+  arity(): number {
+    return 0;
+  }
+
+  call(args: object[]): LoxReturnValue {
+    return new LoxReturnValue(Date.now() / 1000);
+  }
+
+  toString(): string {
+    return "<native fn>";
+  }
+}
